Moved NewCostDistanceMapOp to CostDistanceMapOp
